Keep only one language version of error description
Multi-languagage text options should also be supported in stanza
elements (message/subject, message/body, presence/status). This should
be handled the same way for errors and stanzas, but not complicating
stanzas API too much. The way it was done in ErrorElement was too
complicated. Let's better postpone multi-language support for now, and
do it right later.