Do not abort threads from ThreadPool on every exc.
With other MainLoop implementations it was possible to re-enter
the loop after it raised an exception. This is needed e.g. to gracefully
close a problematic connection. This was not working for ThreadPool
as the thread which raised an exception has already aborted and could
not continue its job.

Because of that the 'test_parse_error' tests from streambase.py took
more time than the others and required a special case when checking the
list of handlers called (DisconnectedEvent was never raised from
ThreadPool, like it was from other MainLoop implementations).

Now the thread exceptions are only queued in the exc_queue, but the
thread keeps on running. This could be probably improved (exception
queue size limit?), so it won't spin crazy on some errors.