SASL API cleanup: no 'encode' option for Reply
sasl.core.Reply constructo does not accept the 'encode' argument any
more. Data is always base64-encoded by .encode(). When the data is None
.encode() returns '', when it is b'', .encode() will return '='.  This
is needed to match the RFC 6120 requirement that: 'If the receiving
entity needs to send additional data of zero length, it MUST transmit
the data as a single equals sign character ("=")'.